pyproject.toml
[tool.poetry]
name = "zm-game-server"
version = "0.1.0"
description = ""
authors = ["Francesco <francesco@nft-factory.club>", "Tommaso <tommaso@nft-factory.club"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = {extras = ["all"], version = "^0.95.2"}
loguru = "^0.7.0"
sentry-sdk = {extras = ["fastapi"], version = "^1.24.0"}
py-healthcheck = "^1.10.1"
tenacity = "^8.2.2"
typer = "^0.9.0"
starlette-exporter = "^0.16.0"
sqlmodel = "^0.0.8"
alembic = "^1.11.1"
aiocache = {extras = ["redis", "memcached"], version = "^0.12.1"}
aioredis = "^2.0.1"
starlette-context = "^0.3.6"
psycopg2 = "^2.9.6"
SQLAlchemy-Utils = "^0.38.3"
six = "^1.16.0"
web3 = "^6.4.0"

[tool.poetry.dev-dependencies]
pytest = "^7.3.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
